+++ src/components/Post/TextArea/ContentArea.jsx
